Refactor PodTemplateAction and NamespaceAction
Leave deprecated methods as they were
