+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/AbstractInvisibleRunAction2.java
@@ -46,19 +46,7 @@ public abstract class AbstractInvisibleRunAction2 extends InvisibleAction implem
 
     protected final Stack<String> stack = new Stack<>();
 
-    private transient Run<?, ?> run;
-
-    public AbstractInvisibleRunAction2() {
-        super();
-    }
-
-    @Deprecated
-    public AbstractInvisibleRunAction2(Run<?, ?> run) {
-        setRun(run);
-    }
-
-    @Deprecated
-    protected abstract AbstractInvisibleRunAction2 createAction(Run<?, ?> run);
+    protected transient Run<?, ?> run;
 
     public Run<?, ?> getRun() {
         return run;
@@ -84,55 +72,6 @@ public abstract class AbstractInvisibleRunAction2 extends InvisibleAction implem
                 bc.commit();
             } catch (InstantiationException | IllegalAccessException e) {
                 throw new RuntimeException("Can not instantiate class " + clazz, e);
-            } finally {
-                bc.abort();
-            }
-        }
-    }
-
-    @Deprecated
-    public void push(String item) throws IOException {
-        if (run == null) {
-            LOGGER.warning("run is null, cannot push");
-            return;
-        }
-        synchronized (run) {
-            BulkChange bc = new BulkChange(run);
-            try {
-                AbstractInvisibleRunAction2 action = run.getAction(this.getClass());
-                if (action == null) {
-                    action = createAction(run);
-                    run.addAction(action);
-                }
-                LOGGER.log(Level.FINEST, "Pushing item {0} to action {1} in run {2}",
-                        new Object[] { item, action, run });
-                action.stack.push(item);
-                bc.commit();
-            } finally {
-                bc.abort();
-            }
-        }
-    }
-
-    @Deprecated
-    public String pop() throws IOException {
-        if (run == null) {
-            LOGGER.warning("run is null, cannot pop");
-            return null;
-        }
-        synchronized (run) {
-            BulkChange bc = new BulkChange(run);
-            try {
-                AbstractInvisibleRunAction2 action = run.getAction(this.getClass());
-                if (action == null) {
-                    action = createAction(run);
-                    run.addAction(action);
-                }
-                String item = action.stack.pop();
-                LOGGER.log(Level.FINEST, "Popped item {0} from action {1} in run {2}",
-                        new Object[] { item, action, run });
-                bc.commit();
-                return item;
             } finally {
                 bc.abort();
             }
